HOTSELLINGPRODUCTS

Approved-by: Ahsan Atiq

+++ database/migrations/2022_09_13_163339_create_products_table.php
@@ -15,12 +15,13 @@ return new class extends Migration
     {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->string('name', 50);
             $table->integer('price');
             $table->timestamps();
+
             $table->foreign('user_id')->references('id')->on('users');
-        });
+        });        
     }
 
     /**
